Move new components to v2 namespace
