+++ osu.Game/Graphics/UserInterfaceV2/LabelledTextBox.cs
@@ -4,10 +4,9 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.UserInterface;
-using osu.Game.Graphics;
 using osu.Game.Graphics.UserInterface;
 
-namespace osu.Game.Screens.Edit.Setup.Components.LabelledComponents
+namespace osu.Game.Graphics.UserInterfaceV2
 {
     public class LabelledTextBox : LabelledComponent<OsuTextBox>
     {
